+++ src/bl4ckscor3/bot/bl4ckb0t/commands/channel/Weather.java
@@ -31,15 +31,15 @@ public class Weather extends BaseChannelCommand
 		
 		try
 		{
-			Document doc = Jsoup.connect("http://api.openweathermap.org/data/2.5/weather?q=" + city.trim() + "&mode=html&APPID=" + Passwords.WEATHERAPIKEY.getPassword()).get();
+			Document doc = Jsoup.connect("http://api.openweathermap.org/data/2.5/weather?q=" + city.trim() + "&mode=xml&APPID=" + Passwords.WEATHERAPIKEY.getPassword()).get();
 
-			Utilities.sendStarMsg(channel, 
-					doc.select("body > div:nth-child(1)").text() + Colors.NORMAL,
-					L10N.getString("w.temperature", channel).replace("#temperature", Colors.BOLD + getTemperature(doc)),
-					L10N.getString("w.humidity", channel).replace("#humidity", Colors.BOLD + doc.select("body > div:nth-child(2) > div:nth-child(3)").text().split(" ")[1]),
-					L10N.getString("w.pressure", channel).replace("#pressure", Colors.BOLD + doc.select("body > div:nth-child(2) > div:nth-child(5)").text().split(" ")[1]),
-					L10N.getString("w.wind", channel).replace("#wind", Colors.BOLD + getWindSpeed(doc)),
-					L10N.getString("w.credit", channel).replace("#link", doc.select("body > div:nth-child(3) > a:nth-child(1)").attr("href")));
+			Utilities.sendStarMsg(channel, doc.select("city").attr("name") + ", " + doc.select("country").text(),
+					L10N.getString("w.temperature", channel).replace("#temperature", Colors.NORMAL + getTemperature(doc)),
+					L10N.getString("w.humidity", channel).replace("#humidity", Colors.NORMAL + doc.select("humidity").attr("value") + doc.select("humidity").attr("unit")),
+					L10N.getString("w.pressure", channel).replace("#pressure", Colors.NORMAL + doc.select("pressure").attr("value") + doc.select("pressure").attr("unit")),
+					L10N.getString("w.wind", channel).replace("#wind", Colors.NORMAL + getWindSpeed(doc)),
+					L10N.getString("w.updated", channel).replace("#date", Colors.NORMAL + doc.select("lastupdate").attr("value").replace("T", " ") + " GMT"),
+					L10N.getString("w.credit", channel).replace("#link", Colors.NORMAL + "http://openweathermap.org/city/" + doc.select("city").attr("id")));
 		}
 		catch(Exception e)
 		{
@@ -50,16 +50,17 @@ public class Weather extends BaseChannelCommand
 
 	private String getTemperature(Document doc)
 	{
-		double celsius = Double.parseDouble(doc.select("body > div:nth-child(2) > div:nth-child(1) > div:nth-child(2) > div:nth-child(1)").text().replace("°C", ""));
-
-		return celsius + "°C | " + Utilities.formatDouble(celsius * (9D / 5D) + 32D) + "°F | " + Utilities.formatDouble(celsius + 273.15D) + "K";
+		double kelvin = Double.parseDouble(doc.select("temperature").attr("value"));
+		String celsius = Utilities.formatDouble(kelvin - 273.15D);
+		
+		return celsius + "°C | " + Utilities.formatDouble(Double.parseDouble(celsius) * (9D / 5D) + 32D) + "°F | " + kelvin + "K";
 	}
 
 	private String getWindSpeed(Document doc)
 	{
-		double ms = Double.parseDouble(doc.select("body > div:nth-child(2) > div:nth-child(4)").text().split(":")[1].trim().replace(" m/s", ""));
+		double ms = Double.parseDouble(doc.select("speed").attr("value"));
 		
-		return ms + " m/s | " + Utilities.formatDouble(ms * 2.2369362920544) + " mph";
+		return ms + " m/s | " + Utilities.formatDouble(ms * 2.2369362920544) + " mph " + (doc.select("direction").attr("code"));
 	}
 	
 	@Override